MOD_5_PORQUE_DEVO_CUIDAR_DO_MEU_CODIGO_PART_2/Ex_Modulo5/ExercicioModulo5/src/main/java/org/example/calculadora/Calculadora.java
package org.example.calculadora;

public class Calculadora {
    private int adicionar;
    private int subtrair;
    private int multiplicar;

    private int dividir;



    public int getAdicionar(int a, int b) {
        return a + b;
    }

    public int getSubtrair(int a, int b) {
        return a - b;
    }

    public int getMultiplicar(int a, int b) {
        return a * b;
    }

    public int getDividir(int a, int b) throws IllegalAccessException {
        if (b == 0){
            throw new IllegalAccessException("Divisão não pode ser por zero");
        }
        return a / b;
    }
}
